chore: move the messaging and fix windows verify builds
